feat: implemented filter

+++ src/main/java/com/pawelnu/projectmanager/utils/PageableResponse.java
@@ -7,5 +7,5 @@ import lombok.Getter;
 @Getter
 public class PageableResponse {
   private long totalElements;
-  private int end;
+  private long end;
 }
